feat Refactors Improve Search
Refactor a lot of old code to make these simpler. Instead of building componments from the three statuses. I decided to just build a single book component.

For Search, all the requests are called on the App.js file to allow for more flexibility in components. Bookcase & SearchBooks can now call from App.js instead of Bookcase originally calling it from SearchBooks, which caused a bit of a confusing mess for myself